% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_curve.qtail.R
\name{calibration_curve.qtail}
\alias{calibration_curve.qtail}
\title{Calibration curve for qtail}
\usage{
calibration_curve.qtail(object, ...)
}
\arguments{
\item{object}{A qtail object}

\item{...}{Additional arguments (not used)}
}
\value{
Data frame with nominal, observed, and count columns
}
\description{
Calibration curve for qtail
}
