% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{check_loss}
\alias{check_loss}
\title{Calculate the check loss.}
\usage{
check_loss(x, tau)
}
\arguments{
\item{x}{A vector.}

\item{tau}{A quantile.}
}
\value{
The average check loss.
}
\description{
Calculate the check loss.
}
