% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.qboost.R
\name{coef.qboost}
\alias{coef.qboost}
\title{Coefficients (feature importance) for \code{qboost}}
\usage{
\method{coef}{qboost}(object, ...)
}
\arguments{
\item{object}{A fitted \code{qboost} model.}

\item{...}{Additional arguments passed to \code{importance.qboost()}.}
}
\value{
A tibble of feature importances.
}
\description{
Coefficients (feature importance) for \code{qboost}
}
