% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.qevt.R
\name{coef.qevt}
\alias{coef.qevt}
\title{Coefficients method for qevt}
\usage{
\method{coef}{qevt}(object, ...)
}
\arguments{
\item{object}{A qevt object}

\item{...}{Additional arguments (not used)}
}
\value{
Named numeric vector of GPD parameters (xi and beta)
}
\description{
Coefficients method for qevt
}
