% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.qtail.R
\name{coef.qtail}
\alias{coef.qtail}
\title{Coefficients method for qtail}
\usage{
\method{coef}{qtail}(object, ...)
}
\arguments{
\item{object}{A qtail object}

\item{...}{Additional arguments (not used)}
}
\value{
Named numeric vector of stacking coefficients
}
\description{
Coefficients method for qtail
}
