% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{compute_stability_score}
\alias{compute_stability_score}
\title{Stability score computation}
\usage{
compute_stability_score(qce, pseudo_r2, overfit_gap)
}
\arguments{
\item{qce}{Quantile coverage error.}

\item{pseudo_r2}{Pseudo-R2 metric.}

\item{overfit_gap}{Overfitting gap (train - CV pinball).}
}
\value{
Stability score roughly bounded between 0 and 1.
}
\description{
Stability score computation
}
\keyword{internal}
