% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{coverage_qce}
\alias{coverage_qce}
\title{Coverage and quantile coverage error}
\usage{
coverage_qce(truth, estimate, tau)
}
\arguments{
\item{truth}{Observed numeric vector.}

\item{estimate}{Predicted numeric vector.}

\item{tau}{Target quantile.}
}
\value{
List with \code{coverage} and \code{qce}.
}
\description{
Coverage and quantile coverage error
}
\keyword{internal}
