% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mqbm.R
\name{.aggregate_calibration}
\alias{.aggregate_calibration}
\title{Aggregate calibration metrics across symbols}
\usage{
.aggregate_calibration(mqbm_object)
}
\arguments{
\item{mqbm_object}{A fitted mqbm model}
}
\value{
List with aggregated calibration metrics
}
\description{
Aggregate calibration metrics across symbols
}
\keyword{internal}
