% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mqbm.R
\name{.aggregate_complexity}
\alias{.aggregate_complexity}
\title{Aggregate complexity metrics across multiplexer values}
\usage{
.aggregate_complexity(mqbm_object)
}
\arguments{
\item{mqbm_object}{A fitted mqbm model}
}
\value{
List with complexity metrics
}
\description{
Aggregate complexity metrics across multiplexer values
}
\keyword{internal}
