% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mqbm.R
\name{.aggregate_symbol_metrics}
\alias{.aggregate_symbol_metrics}
\title{Aggregate metrics across group-specific models}
\usage{
.aggregate_symbol_metrics(mqbm_object)
}
\arguments{
\item{mqbm_object}{A fitted mqbm model}
}
\value{
List with aggregated metrics for training and validation
}
\description{
Aggregate metrics across group-specific models
}
\keyword{internal}
