% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mqbm.R
\name{.aggregate_value_metrics}
\alias{.aggregate_value_metrics}
\title{Aggregate metrics across multiplexer values}
\usage{
.aggregate_value_metrics(mqbm_object)
}
\arguments{
\item{mqbm_object}{A fitted mqbm model}
}
\value{
List with aggregated metrics for training and validation
}
\description{
Aggregate metrics across multiplexer values
}
\keyword{internal}
