% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mqbm.R
\name{.compute_symbol_table}
\alias{.compute_symbol_table}
\title{Create per-symbol summary table}
\usage{
.compute_symbol_table(mqbm_object, metrics_list, calibration_list)
}
\arguments{
\item{mqbm_object}{A fitted mqbm model}

\item{metrics_list}{Output from .aggregate_symbol_metrics}

\item{calibration_list}{Output from .aggregate_calibration}
}
\value{
Tibble with per-symbol summary
}
\description{
Create per-symbol summary table
}
\keyword{internal}
