% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{extract_tree_leaves}
\alias{extract_tree_leaves}
\title{Extract leaf count from a single tree}
\usage{
extract_tree_leaves(tree)
}
\description{
Extract leaf count from a single tree
}
\keyword{internal}
