% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotone_quantiles.R
\name{fit_quantile_models}
\alias{fit_quantile_models}
\title{Fit independent quantile models}
\usage{
fit_quantile_models(x, y, taus, params = list(), ...)
}
\arguments{
\item{x}{Feature matrix/data.frame}

\item{y}{Numeric target}

\item{taus}{Numeric vector of quantile levels}

\item{params}{List of LightGBM parameters (per model)}

\item{...}{Extra args to qboost}
}
\value{
Named list of models keyed by tau
}
\description{
Fit independent quantile models
}
