% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.mqtail.R
\name{fitted.mqtail}
\alias{fitted.mqtail}
\title{Fitted values from mqtail model}
\usage{
\method{fitted}{mqtail}(object, type = c("surface", "quantile"), ...)
}
\arguments{
\item{object}{A fitted mqtail model}

\item{type}{Either "surface" (default) or "quantile"}

\item{...}{Additional arguments (not used)}
}
\description{
Fitted values from mqtail model
}
