% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.qtail.R
\name{fitted.qtail}
\alias{fitted.qtail}
\title{Fitted values method for qtail}
\usage{
\method{fitted}{qtail}(object, ...)
}
\arguments{
\item{object}{A qtail object}

\item{...}{Additional arguments (not used)}
}
\value{
Numeric vector of fitted values
}
\description{
Fitted values method for qtail
}
