% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{gain_per_leaf_metric}
\alias{gain_per_leaf_metric}
\title{Gain per leaf metric}
\usage{
gain_per_leaf_metric(importance_df, total_leaves)
}
\arguments{
\item{importance_df}{Importance tibble.}

\item{total_leaves}{Total number of leaves.}
}
\value{
Gain per leaf.
}
\description{
Gain per leaf metric
}
\keyword{internal}
