% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{get_default_leaves}
\alias{get_default_leaves}
\title{Get default leaves parameter from model}
\usage{
get_default_leaves(model)
}
\description{
Get default leaves parameter from model
}
\keyword{internal}
