% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{grad_check_loss}
\alias{grad_check_loss}
\title{Calculate the gradient of the check loss.}
\usage{
grad_check_loss(x, tau)
}
\arguments{
\item{x}{A vector.}

\item{tau}{A quantile.}
}
\value{
The gradient of the check loss.
}
\description{
Calculate the gradient of the check loss.
}
