% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.qbm.R
\name{importance.qbm}
\alias{importance.qbm}
\title{Feature importance for a \code{qbm} model}
\usage{
\method{importance}{qbm}(object, ...)
}
\arguments{
\item{object}{A fitted \code{qbm} model.}

\item{...}{Additional arguments passed to \code{lightgbm::lgb.importance()}.}
}
\value{
A tibble of feature importances with columns \code{feature}, \code{gain}, \code{cover}, \code{freq}.
}
\description{
Feature importance for a \code{qbm} model
}
