% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.qboost.R
\name{importance.qboost}
\alias{importance.qboost}
\title{Feature importance for a \code{qboost} model}
\usage{
importance.qboost(object, ...)
}
\arguments{
\item{object}{A fitted \code{qboost} model.}

\item{...}{Additional arguments passed to \code{lightgbm::lgb.importance()}.}
}
\value{
A tibble of feature importances with columns \code{feature}, \code{gain}, \code{cover}, \code{freq}.
}
\description{
Feature importance for a \code{qboost} model
}
