% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.qbm.R
\name{importance}
\alias{importance}
\title{Extract feature importance from a model}
\usage{
importance(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A tibble of feature importances.
}
\description{
Generic function for extracting feature importance from fitted models.
}
