% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{importance_entropy}
\alias{importance_entropy}
\title{Importance entropy (Shannon entropy of normalized gains)}
\usage{
importance_entropy(importance_df)
}
\arguments{
\item{importance_df}{Importance tibble.}
}
\value{
Entropy value.
}
\description{
Importance entropy (Shannon entropy of normalized gains)
}
\keyword{internal}
