% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{leaf_stats}
\alias{leaf_stats}
\title{Extract leaves statistics from LightGBM model}
\usage{
leaf_stats(model, best_iter = NULL)
}
\arguments{
\item{model}{An \code{lgb.Booster} object.}
}
\value{
List with total leaves and average leaves per tree.
}
\description{
Extract leaves statistics from LightGBM model
}
\keyword{internal}
