% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(truth, estimate)
}
\arguments{
\item{truth}{Observed numeric vector.}

\item{estimate}{Predicted numeric vector.}
}
\value{
Scalar MAE.
}
\description{
Mean absolute error
}
\keyword{internal}
