% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{pinball_loss}
\alias{pinball_loss}
\title{Pinball loss for quantile regression}
\usage{
pinball_loss(truth, estimate, tau)
}
\arguments{
\item{truth}{Observed numeric vector.}

\item{estimate}{Predicted numeric vector.}

\item{tau}{Target quantile.}
}
\value{
Numeric vector of losses.
}
\description{
Pinball loss for quantile regression
}
\keyword{internal}
