% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{pinball_loss_mean}
\alias{pinball_loss_mean}
\title{Mean pinball loss}
\usage{
pinball_loss_mean(truth, estimate, tau)
}
\arguments{
\item{truth}{Observed numeric vector.}

\item{estimate}{Predicted numeric vector.}

\item{tau}{Target quantile.}
}
\value{
Scalar mean loss.
}
\description{
Mean pinball loss
}
\keyword{internal}
