% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{pinball_plot_data}
\alias{pinball_plot_data}
\title{Data for pinball loss plot}
\usage{
pinball_plot_data(truth, estimate, tau)
}
\arguments{
\item{truth}{Observed numeric vector.}

\item{estimate}{Predicted numeric vector.}

\item{tau}{Target quantile.}
}
\value{
Data frame with residuals and loss.
}
\description{
Data for pinball loss plot
}
\keyword{internal}
