% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit_qevt.R
\name{pit_plot_qevt}
\alias{pit_plot_qevt}
\title{PIT Plot for qevt}
\usage{
pit_plot_qevt(object, breaks = 40, ...)
}
\arguments{
\item{object}{A qevt object}

\item{breaks}{Number of histogram breaks (default 40)}

\item{...}{Additional arguments passed to hist}
}
\value{
NULL (invisibly)
}
\description{
PIT Plot for qevt
}
\keyword{internal}
