% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit_qevt.R
\name{pit_qevt}
\alias{pit_qevt}
\title{Probability Integral Transform for qevt}
\usage{
pit_qevt(object, ...)
}
\arguments{
\item{object}{A qevt object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of PIT values
}
\description{
Probability Integral Transform for qevt
}
\keyword{internal}
