% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.qevt.R
\name{plot.qevt}
\alias{plot.qevt}
\title{Plot diagnostics for qevt}
\usage{
\method{plot}{qevt}(x, newdata = NULL, obs = 1, ...)
}
\arguments{
\item{x}{qevt model}

\item{newdata}{Feature matrix/data.frame; defaults to training data stored in model}

\item{obs}{integer row index to plot (default 1)}

\item{...}{unused}
}
\description{
Plot diagnostics for qevt
}
