% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.qtail.R
\name{plot.qtail}
\alias{plot.qtail}
\title{Plot method for qtail}
\usage{
\method{plot}{qtail}(
  x,
  type = c("calibration", "stack_coeffs", "evt_tail", "gpd_qq", "residuals"),
  ...
)
}
\arguments{
\item{x}{A qtail object}

\item{type}{Type of plot to produce}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
NULL (invisibly)
}
\description{
Plot method for qtail
}
