% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{plot_calibration_curve}
\alias{plot_calibration_curve}
\title{Calibration curve plot}
\usage{
plot_calibration_curve(df)
}
\arguments{
\item{df}{Data frame from \code{calibration_curve()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Calibration curve plot
}
\keyword{internal}
