% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{plot_feature_importance}
\alias{plot_feature_importance}
\title{Feature importance plot}
\usage{
plot_feature_importance(df, top_n = 25)
}
\arguments{
\item{df}{Data frame returned by \code{importance.qbm()}.}

\item{top_n}{Maximum number of features to display.}
}
\value{
A \code{ggplot} object.
}
\description{
Feature importance plot
}
\keyword{internal}
