% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{plot_pinball_loss}
\alias{plot_pinball_loss}
\title{Pinball loss plot}
\usage{
plot_pinball_loss(df)
}
\arguments{
\item{df}{Data frame from \code{pinball_plot_data()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Pinball loss plot
}
\keyword{internal}
