% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{plot_qq_quantiles}
\alias{plot_qq_quantiles}
\title{QQ plot between predicted and empirical quantiles}
\usage{
plot_qq_quantiles(df)
}
\arguments{
\item{df}{Data frame from \code{qq_plot_data()}.}
}
\value{
A \code{ggplot} object.
}
\description{
QQ plot between predicted and empirical quantiles
}
\keyword{internal}
