% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mqbm.R
\name{predict.mqbm}
\alias{predict.mqbm}
\title{Predict from a \code{mqbm} model}
\usage{
\method{predict}{mqbm}(object, newdata, symbol = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{mqbm} object.}

\item{newdata}{New data.frame or matrix of predictors. For models fit with a
formula, a data.frame with the same predictor columns used at training
time is expected. Must contain a \code{symbol} column OR provide \code{symbol} as a
separate argument.}

\item{symbol}{Optional vector of symbols for prediction. If not provided,
will look for a \code{symbol} column in \code{newdata}.}

\item{...}{Additional arguments passed to \code{stats::predict()}.}
}
\value{
Numeric vector of predicted quantiles.
}
\description{
Predict from a \code{mqbm} model
}
