% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.qboost.R
\name{predict.qboost}
\alias{predict.qboost}
\title{Predict from a \code{qboost} model}
\usage{
\method{predict}{qboost}(object, newdata, ...)
}
\arguments{
\item{object}{A fitted \code{qboost} object.}

\item{newdata}{New data.frame or matrix of predictors.}

\item{...}{Additional arguments passed to \code{stats::predict()}.}
}
\value{
Numeric vector of predicted quantiles.
}
\description{
Predict from a \code{qboost} model
}
