% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotone_quantiles.R
\name{predict_quantiles_raw}
\alias{predict_quantiles_raw}
\title{Raw predictions for all taus}
\usage{
predict_quantiles_raw(models, newdata, taus)
}
\arguments{
\item{models}{List from fit_quantile_models}

\item{newdata}{Feature matrix/data.frame}

\item{taus}{Quantile levels (in same order used for fitting)}
}
\value{
Matrix n x K of raw quantile predictions
}
\description{
Raw predictions for all taus
}
