% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mqbm.R
\name{print.mqbm}
\alias{print.mqbm}
\title{Print method for \code{mqbm}}
\usage{
\method{print}{mqbm}(x, ...)
}
\arguments{
\item{x}{A \code{mqbm} object.}

\item{...}{Additional arguments (unused).}
}
\value{
The input object, invisibly.
}
\description{
Print method for \code{mqbm}
}
