% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mqbm.R
\name{print.mqbm_summary}
\alias{print.mqbm_summary}
\title{Print method for mqbm_summary}
\usage{
\method{print}{mqbm_summary}(x, ...)
}
\arguments{
\item{x}{An mqbm_summary object}

\item{...}{Additional arguments (unused)}
}
\value{
The input object, invisibly
}
\description{
Print method for mqbm_summary
}
