% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mqtail.R
\name{print.mqtail}
\alias{print.mqtail}
\title{Print method for mqtail``}
\usage{
\method{print}{mqtail}(x, ...)
}
\arguments{
\item{x}{A mqtail object.}

\item{...}{Additional arguments (unused).}
}
\value{
The input object, invisibly.
}
\description{
Print method for mqtail``
}
