% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mqtail.R
\name{print.mqtail_summary}
\alias{print.mqtail_summary}
\title{Print method for mqtail_summary}
\usage{
\method{print}{mqtail_summary}(x, ...)
}
\arguments{
\item{x}{An mqtail_summary object}

\item{...}{Additional arguments (unused)}
}
\value{
The input object, invisibly
}
\description{
Print method for mqtail_summary
}
