% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qbm.R
\name{print.qbm_summary}
\alias{print.qbm_summary}
\title{Print method for \code{qbm_summary}}
\usage{
\method{print}{qbm_summary}(x, ...)
}
\arguments{
\item{x}{A \code{qbm_summary} object.}

\item{...}{Additional arguments (unused).}
}
\value{
The input object, invisibly.
}
\description{
Print method for \code{qbm_summary}
}
