% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qboost.R
\name{print.qboost_summary}
\alias{print.qboost_summary}
\title{Print method for \code{qboost_summary}}
\usage{
\method{print}{qboost_summary}(x, ...)
}
\arguments{
\item{x}{A \code{qboost_summary} object.}

\item{...}{Additional arguments (unused).}
}
\value{
The input object, invisibly.
}
\description{
Print method for \code{qboost_summary}
}
