% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qevt.R
\name{print.qevt_summary}
\alias{print.qevt_summary}
\title{Print method for qevt_summary}
\usage{
\method{print}{qevt_summary}(x, ...)
}
\arguments{
\item{x}{A qevt_summary object}

\item{...}{Additional arguments (unused)}
}
\value{
The input object, invisibly
}
\description{
Print method for qevt_summary
}
