% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.qtail.R
\name{print.qtail}
\alias{print.qtail}
\title{Print method for qtail}
\usage{
\method{print}{qtail}(x, ...)
}
\arguments{
\item{x}{A qtail object}

\item{...}{Additional arguments passed to summary}
}
\value{
The object (invisibly)
}
\description{
Print method for qtail
}
