% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qtail.R
\name{print.qtail_summary}
\alias{print.qtail_summary}
\title{Print method for qtail_summary}
\usage{
\method{print}{qtail_summary}(x, ...)
}
\arguments{
\item{x}{A qtail_summary object}

\item{...}{Additional arguments (unused)}
}
\value{
The input object, invisibly
}
\description{
Print method for qtail_summary
}
