% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{qq_plot_data}
\alias{qq_plot_data}
\title{QQ plot data between predicted quantile and empirical distribution}
\usage{
qq_plot_data(truth, estimate, probs = seq(0.05, 0.95, by = 0.05))
}
\arguments{
\item{truth}{Observed numeric vector.}

\item{estimate}{Predicted numeric vector.}

\item{probs}{Probability grid.}
}
\value{
Data frame with empirical and predicted quantiles.
}
\description{
QQ plot data between predicted quantile and empirical distribution
}
\keyword{internal}
