% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit_qevt.R
\name{qq_plot_qevt}
\alias{qq_plot_qevt}
\title{QQ Plot for qevt}
\usage{
qq_plot_qevt(object, probs = seq(0.05, 0.95, by = 0.05), ...)
}
\arguments{
\item{object}{A qevt object}

\item{probs}{Probability grid for QQ plot (default seq(0.05, 0.95, 0.05))}

\item{...}{Additional arguments (not used)}
}
\value{
NULL (invisibly)
}
\description{
QQ Plot for qevt
}
\keyword{internal}
