% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtail_calibration_slope.R
\name{qtail_calibration_slope}
\alias{qtail_calibration_slope}
\title{Quantile Calibration Slope for qtail}
\usage{
qtail_calibration_slope(object, tau = object$tau_target, ...)
}
\arguments{
\item{object}{A qtail object}

\item{tau}{Quantile level to evaluate (default: object$tau_target)}

\item{...}{Additional arguments (not used)}
}
\value{
List with slope, intercept, and tau
}
\description{
Quantile Calibration Slope for qtail
}
