% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtail_pit.R
\name{qtail_pit}
\alias{qtail_pit}
\title{Probability Integral Transform for qtail}
\usage{
qtail_pit(object, ...)
}
\arguments{
\item{object}{A qtail object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of PIT values
}
\description{
Probability Integral Transform for qtail
}
