% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtail_pit_plot.R
\name{qtail_pit_plot}
\alias{qtail_pit_plot}
\title{PIT Plot for qtail}
\usage{
qtail_pit_plot(object, breaks = 40, ...)
}
\arguments{
\item{object}{A qtail object}

\item{breaks}{Number of histogram breaks}

\item{...}{Additional arguments passed to hist}
}
\value{
NULL (invisibly)
}
\description{
PIT Plot for qtail
}
