% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtail_qce.R
\name{qtail_qce}
\alias{qtail_qce}
\title{Quantile Calibration Error for qtail}
\usage{
qtail_qce(object, taus = NULL, ...)
}
\arguments{
\item{object}{A qtail object}

\item{taus}{Optional vector of quantile levels to evaluate}

\item{...}{Additional arguments (not used)}
}
\value{
Data frame with tau, observed, and qce columns
}
\description{
Quantile Calibration Error for qtail
}
