% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtail_stability.R
\name{qtail_stability}
\alias{qtail_stability}
\title{Stability Metric for qtail}
\usage{
qtail_stability(object, ...)
}
\arguments{
\item{object}{A qtail object}

\item{...}{Additional arguments (not used)}
}
\value{
List with stability score and components
}
\description{
Stability Metric for qtail
}
