% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtail_validate.R
\name{qtail_validate}
\alias{qtail_validate}
\title{Out-of-sample Validation for qtail}
\usage{
qtail_validate(object, x_new, y_new, taus = NULL, ...)
}
\arguments{
\item{object}{A qtail object}

\item{x_new}{New feature matrix}

\item{y_new}{New response vector}

\item{taus}{Optional vector of quantile levels to evaluate}

\item{...}{Additional arguments (not used)}
}
\value{
Data frame with validation metrics
}
\description{
Out-of-sample Validation for qtail
}
