% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.qtail.R
\name{residuals.qtail}
\alias{residuals.qtail}
\title{Residuals method for qtail}
\usage{
\method{residuals}{qtail}(object, ...)
}
\arguments{
\item{object}{A qtail object}

\item{...}{Additional arguments (not used)}
}
\value{
Numeric vector of residuals
}
\description{
Residuals method for qtail
}
