% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{smooth_check_loss}
\alias{smooth_check_loss}
\title{Calculate the smoothed check loss.}
\usage{
smooth_check_loss(x, tau, h = 0.1, kernel = "Gaussian")
}
\arguments{
\item{x}{A vector.}

\item{tau}{A quantile.}

\item{h}{The bandwith for smoothing.}

\item{kernel}{The kernel.}
}
\value{
The average smoothed check loss.
}
\description{
Calculate the smoothed check loss.
}
