% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotone_quantiles.R
\name{stack_quantiles}
\alias{stack_quantiles}
\title{Stack monotone quantiles with ridge}
\usage{
stack_quantiles(preds_mono, y, lambda = 0.01)
}
\arguments{
\item{preds_mono}{n x K matrix of monotone predictions}

\item{y}{Numeric target for stacking fit}

\item{lambda}{Ridge penalty}
}
\value{
List with fit and coefficients
}
\description{
Stack monotone quantiles with ridge
}
