% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qboost.R
\name{summary.qboost}
\alias{summary.qboost}
\title{Summarise a \code{qboost} model}
\usage{
\method{summary}{qboost}(object, ...)
}
\arguments{
\item{object}{A fitted \code{qboost} object.}

\item{...}{Unused, included for compatibility.}
}
\value{
An object of class \code{qboost_summary}.
}
\description{
Summarise a \code{qboost} model
}
