% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_params.qtail.R
\name{tail_params.qtail}
\alias{tail_params.qtail}
\title{Tail parameters method for qtail}
\usage{
tail_params.qtail(object, ...)
}
\arguments{
\item{object}{A qtail object}

\item{...}{Additional arguments (not used)}
}
\value{
List of EVT parameters
}
\description{
Tail parameters method for qtail
}
