% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metrics.R
\name{tidy_importance}
\alias{tidy_importance}
\title{Tidy LightGBM feature importance}
\usage{
tidy_importance(model, ...)
}
\arguments{
\item{model}{An \code{lgb.Booster} object.}

\item{...}{Additional arguments to \code{lightgbm::lgb.importance()}.}
}
\value{
Tibble with feature, gain, cover, freq.
}
\description{
Tidy LightGBM feature importance
}
\keyword{internal}
