% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_selection_step.R
\name{update_selection_step}
\alias{update_selection_step}
\title{Greedy procedure to determine the selection covariate}
\usage{
update_selection_step(X, residuals, tau, h = 0.1, kernel = "Gaussian")
}
\arguments{
\item{X}{A matrix of covariates.}

\item{residuals}{A vector of residuals.}

\item{tau}{A quantile.}

\item{h}{The bandwith for smoothing.}

\item{kernel}{The kernel.}
}
\value{
A list with components
\item{sel_cov}{The index of the selected covariates.}
\item{cor}{The corresponding correlation.}
}
\description{
Calculates the gradient of the smoothed check-loss for the residuals
and selects the column of X which has the highest absolute correlation.
}
