% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.qbm.R
\name{coef.qbm}
\alias{coef.qbm}
\title{Coefficients (feature importance) for \code{qbm}}
\usage{
\method{coef}{qbm}(object, ...)
}
\arguments{
\item{object}{A fitted \code{qbm} model.}

\item{...}{Additional arguments passed to \code{importance.qbm()}.}
}
\value{
A tibble of feature importances.
}
\description{
Coefficients (feature importance) for \code{qbm}
}
