% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.qtail.R
\name{plot.qtail}
\alias{plot.qtail}
\title{Plot method for qtail}
\usage{
\method{plot}{qtail}(
  x,
  type = c("all", "calibration", "stack_coeffs", "evt_tail", "gpd_qq", "residuals",
    "coverage"),
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{A qtail object}

\item{type}{Type of plot to produce. If "all" (default), produces all diagnostic plots.}

\item{plot}{Logical; if TRUE, plots are printed. If FALSE, returns list of plots.}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
If plot=FALSE, a list of ggplot objects. Otherwise NULL (invisibly).
}
\description{
Plot method for qtail
}
