% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mqtail.R
\name{predict.mqtail}
\alias{predict.mqtail}
\title{Predict from mqtail model}
\usage{
\method{predict}{mqtail}(object, newdata, type = c("surface", "quantile"), ...)
}
\arguments{
\item{object}{A fitted mqtail model}

\item{newdata}{Data for prediction}

\item{type}{Either "surface" (default) or "quantile"}

\item{...}{Passed to predict.mqbm}
}
\description{
Predict from mqtail model
}
