% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.qevt.R
\name{predict.qevt}
\alias{predict.qevt}
\title{Predict method for qevt}
\usage{
\method{predict}{qevt}(object, newdata, ...)
}
\arguments{
\item{object}{qevt model}

\item{newdata}{Feature matrix/data.frame}

\item{...}{unused}
}
\value{
list(raw, monotone, taus)
}
\description{
Predict method for qevt
}
