% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotone_quantiles.R
\name{predict_quantiles_monotone}
\alias{predict_quantiles_monotone}
\title{Convenience: predict and enforce monotone}
\usage{
predict_quantiles_monotone(models, newdata, taus)
}
\description{
Convenience: predict and enforce monotone
}
