% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qboost.R
\name{print.qboost}
\alias{print.qboost}
\title{Print method for \code{qboost}}
\usage{
\method{print}{qboost}(x, ...)
}
\arguments{
\item{x}{A \code{qboost} object.}

\item{...}{Additional arguments (unused).}
}
\value{
The input object, invisibly.
}
\description{
Print method for \code{qboost}
}
