% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qbm.R
\name{print.qbm}
\alias{print.qbm}
\title{Print method for \code{qbm}}
\usage{
\method{print}{qbm}(x, ...)
}
\arguments{
\item{x}{A \code{qbm} object.}

\item{...}{Additional arguments (unused).}
}
\value{
The input object, invisibly.
}
\description{
Print method for \code{qbm}
}
