% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qboost.R
\name{qboost}
\alias{qboost}
\title{Quantile Boosting for high-dimensional data}
\usage{
qboost(
  X,
  Y,
  tau = 0.5,
  m_stop = 50,
  h = 0.1,
  kernel = "Gaussian",
  stepsize = NULL
)
}
\arguments{
\item{X}{(\code{matrix}) \cr
A matrix of covariates.}

\item{Y}{(\code{numeric()}) \cr
A vector of responses.}

\item{tau}{(\code{numeric(1L)}) \cr
A quantile.}

\item{m_stop}{(\code{integer(1L)}) \cr
The number of boosting steps.}

\item{h}{(\code{numeric(1L)}) \cr
The bandwith for smoothing.}

\item{kernel}{(\code{character(1L)}) \cr
The kernel for smoothing.}

\item{stepsize}{(\code{numeric(1L)}) \cr
The stepsize of the boosting procedure. Set to \code{NULL} for WCGA.}
}
\value{
An object with S3 class \code{qboost}
\itemize{
\item \code{coeff_path} The cofficients along the boosting steps as a matrix (starts with a zero vector).
\item \code{selection_path} The selected covariates as a vector.
}
}
\description{
Applying a greedy algorithm to minimize the expected smoothed check loss.
Employs the Weak Restricted Greedy Algorithm. For the orthogonal variant, set \code{stepsize} to \code{NULL}.
}
\seealso{
\code{predict.qboost}
}
