% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qboost.R
\name{qboost}
\alias{qboost}
\title{Title}
\usage{
qboost(
  X,
  Y,
  tau = 0.5,
  m_stop = 50,
  h = 0.1,
  kernel = "Gaussian",
  stepsize = NULL
)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{Y}{A vector of responses.}

\item{tau}{A quantile.}

\item{m_stop}{The number of boosting steps.}

\item{h}{The bandwith for smoothing.}

\item{kernel}{The kernel for smoothing.}

\item{stepsize}{The stepsize of the boosting procedure.}
}
\value{
A list with components
\item{coeff_path}{The cofficients along the boosting steps as a matrix (starts with a zero vector).}
\item{selection_path}{The selected covariates as a vector.}
}
\description{
Title
}
