% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qevt.R
\name{qevt}
\alias{qevt}
\title{Fit extreme quantile model with EVT and PAVA monotonicity}
\usage{
qevt(
  X,
  y,
  tau_target = 0.9999,
  tau0 = 0.997,
  tau1 = 0.9985,
  tau2 = 0.999,
  tau3 = 0.9993
)
}
\arguments{
\item{X}{Feature matrix/data.frame}

\item{y}{Numeric response}

\item{tau_target}{Final extreme tau (default 0.9999)}

\item{tau0}{Threshold quantile for EVT (default 0.997)}

\item{tau1}{Intermediate EVT quantile (default 0.9985)}

\item{tau2}{Intermediate EVT quantile (default 0.999)}

\item{tau3}{Intermediate EVT quantile (default 0.9993)}
}
\value{
Object of class 'qevt'
}
\description{
Fit extreme quantile model with EVT and PAVA monotonicity
}
