% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtail.R
\name{qtail}
\alias{qtail}
\title{Fit qtail model}
\usage{
qtail(
  x,
  y,
  taus = NULL,
  tail = c("upper", "lower"),
  threshold_tau = NULL,
  params = list(),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Feature matrix}

\item{y}{Target vector}

\item{taus}{Quantile levels to fit}

\item{tail}{Tail type: "upper" or "lower"}

\item{threshold_tau}{Threshold quantile for EVT}

\item{params}{Parameters passed to fit_qboost}

\item{verbose}{Logical; if TRUE, print progress messages}

\item{...}{Additional arguments passed to fit_qboost}
}
\value{
An object of class "qtail"
}
\description{
Fit qtail model
}
