% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{smooth_check_loss}
\alias{smooth_check_loss}
\title{Smoothed check loss}
\usage{
smooth_check_loss(x, tau, h = 0.1, kernel = "Gaussian")
}
\arguments{
\item{x}{(\code{numeric()}) \cr
A vector.}

\item{tau}{(\code{numeric(1L)}) \cr
A quantile.}

\item{h}{(\code{numeric(1L)}) \cr
The bandwith for smoothing.}

\item{kernel}{(\code{character(1L)}) \cr
The kernel for smoothing.}
}
\value{
The average smoothed check loss.
}
\description{
Calculate the smoothed check loss.
}
