% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_selection_step.R
\name{update_selection_step}
\alias{update_selection_step}
\title{Greedy procedure to determine the selection covariate}
\usage{
update_selection_step(X, residuals, tau, h = 0.1, kernel = "Gaussian")
}
\arguments{
\item{X}{(\code{matrix}) \cr
A matrix of covariates.}

\item{residuals}{(\code{numeric()}) \cr
A vector of residuals.}

\item{tau}{(\code{numeric(1L)}) \cr
A quantile.}

\item{h}{(\code{numeric(1L)}) \cr
The bandwith for smoothing.}

\item{kernel}{(\code{character(1L)}) \cr
The kernel for smoothing.}
}
\value{
A list with components
\itemize{
\item \code{sel_cov} The index of the selected covariates.
\item \code{cor} The corresponding correlation.
}
}
\description{
Calculates the gradient of the smoothed check-loss for the residuals
and selects the column of X which has the highest absolute correlation.
}
